feat graphics: improve archi with menu and game
