+++ client/include/gui/states/PrivateServerLobbyState.hpp
@@ -1,15 +1,15 @@
-#ifndef CLIENT_PRIVATE_SERVER_STATE_HPP
-#define CLIENT_PRIVATE_SERVER_STATE_HPP
+#ifndef CLIENT_PRIVATE_SERVER_LOBBY_STATE_HPP
+#define CLIENT_PRIVATE_SERVER_LOBBY_STATE_HPP
 
 #include "State.hpp"
-#include "StateManager.hpp"
+#include "../StateManager.hpp"
 #include <SFML/Graphics.hpp>
 #include <string>
 
 namespace rtype::client::gui {
-    class PrivateServerState : public State {
+    class PrivateServerLobbyState : public State {
     public:
-        PrivateServerState(StateManager& stateManager, const std::string& username);
+        PrivateServerLobbyState(StateManager& stateManager, const std::string& username, const std::string& serverCode, bool isAdmin);
         
         void handleEvent(const sf::Event& event) override;
         void update(float deltaTime) override;
@@ -20,35 +20,32 @@ namespace rtype::client::gui {
     private:
         StateManager& stateManager;
         std::string username;
+        std::string serverCode;
+        bool isAdmin;
         sf::Font font;
         
         // UI Elements
-        sf::Text titleText;
-        sf::Text serverCodeText;
-        sf::Text serverCodeHintText;
-        sf::RectangleShape serverCodeBox;
-        sf::Text joinButton;
-        sf::RectangleShape joinButtonRect;
-        sf::Text createButton;
-        sf::RectangleShape createButtonRect;
+        sf::Text playersReadyText;
+        sf::Text actionButton; // "Ready" for players, "Start" for admin
+        sf::RectangleShape actionButtonRect;
         sf::Text returnButton;
         sf::RectangleShape returnButtonRect;
+        sf::Text serverCodeDisplay;
         
         // State
-        std::string serverCode;
-        bool isTyping;
-        float cursorTimer;
-        bool showCursor;
+        bool isReady;
+        int playersReady;
         
         // Helper methods
         void setupUI();
         void updateLayout(const sf::Vector2u& windowSize);
         bool isPointInRect(const sf::Vector2f& point, const sf::RectangleShape& rect);
         void centerText(sf::Text& text, float x, float y);
-        void joinServer();
-        void createServer();
-        bool isValidServerCode(const std::string& code);
+        void toggleReady();
+        void startGame();
+        void updatePlayersReadyText();
+        void updateActionButton();
     };
 }
 
-#endif // CLIENT_PRIVATE_SERVER_STATE_HPP
+#endif // CLIENT_PRIVATE_SERVER_LOBBY_STATE_HPP
