+++ client/include/gui/states/PrivateServerState.hpp
@@ -1,15 +1,15 @@
-#ifndef CLIENT_PUBLIC_SERVER_STATE_HPP
-#define CLIENT_PUBLIC_SERVER_STATE_HPP
+#ifndef CLIENT_PRIVATE_SERVER_STATE_HPP
+#define CLIENT_PRIVATE_SERVER_STATE_HPP
 
 #include "State.hpp"
-#include "StateManager.hpp"
+#include "../StateManager.hpp"
 #include <SFML/Graphics.hpp>
 #include <string>
 
 namespace rtype::client::gui {
-    class PublicServerState : public State {
+    class PrivateServerState : public State {
     public:
-        PublicServerState(StateManager& stateManager, const std::string& username);
+        PrivateServerState(StateManager& stateManager, const std::string& username);
         
         void handleEvent(const sf::Event& event) override;
         void update(float deltaTime) override;
@@ -23,24 +23,32 @@ namespace rtype::client::gui {
         sf::Font font;
         
         // UI Elements
-        sf::Text playersReadyText;
-        sf::Text readyButton;
-        sf::RectangleShape readyButtonRect;
+        sf::Text titleText;
+        sf::Text serverCodeText;
+        sf::Text serverCodeHintText;
+        sf::RectangleShape serverCodeBox;
+        sf::Text joinButton;
+        sf::RectangleShape joinButtonRect;
+        sf::Text createButton;
+        sf::RectangleShape createButtonRect;
         sf::Text returnButton;
         sf::RectangleShape returnButtonRect;
         
         // State
-        bool isReady;
-        int playersReady;
+        std::string serverCode;
+        bool isTyping;
+        float cursorTimer;
+        bool showCursor;
         
         // Helper methods
         void setupUI();
         void updateLayout(const sf::Vector2u& windowSize);
         bool isPointInRect(const sf::Vector2f& point, const sf::RectangleShape& rect);
         void centerText(sf::Text& text, float x, float y);
-        void toggleReady();
-        void updatePlayersReadyText();
+        void joinServer();
+        void createServer();
+        bool isValidServerCode(const std::string& code);
     };
 }
 
-#endif // CLIENT_PUBLIC_SERVER_STATE_HPP
+#endif // CLIENT_PRIVATE_SERVER_STATE_HPP
