+++ client/src/gui/states/MainMenuState.cpp
@@ -1,9 +1,12 @@
-#include "MainMenu.hpp"
+#include "gui/states/MainMenuState.hpp"
+#include "gui/states/PublicServerState.hpp"
+#include "gui/states/PrivateServerState.hpp"
 #include <iostream>
+#include <cstdlib>
 
 namespace rtype::client::gui {
-    MainMenu::MainMenu(sf::RenderWindow& window)
-        : window(window), isTyping(false), cursorTimer(0.0f), showCursor(true) {
+    MainMenuState::MainMenuState(StateManager& stateManager)
+        : stateManager(stateManager), isTyping(false), cursorTimer(0.0f), showCursor(true) {
         
         // Try to load fonts in order of preference
         bool fontLoaded = false;
@@ -31,10 +34,9 @@ namespace rtype::client::gui {
         }
         
         setupUI();
-        updateLayout();
     }
     
-    void MainMenu::setupUI() {
+    void MainMenuState::setupUI() {
         // Title setup
         titleText.setFont(font);
         titleText.setString("THE TOP R-TYPE");
@@ -77,8 +79,11 @@ namespace rtype::client::gui {
         privateButtonRect.setOutlineThickness(2.0f);
     }
     
-    void MainMenu::updateLayout() {
-        sf::Vector2u windowSize = window.getSize();
+    void MainMenuState::onEnter() {
+        // Called when this state becomes active
+    }
+    
+    void MainMenuState::updateLayout(const sf::Vector2u& windowSize) {
         float centerX = windowSize.x / 2.0f;
         float centerY = windowSize.y / 2.0f;
         
@@ -117,9 +122,9 @@ namespace rtype::client::gui {
                   privateButtonRect.getPosition().y + buttonHeight / 2);
     }
     
-    void MainMenu::handleEvent(const sf::Event& event) {
+    void MainMenuState::handleEvent(const sf::Event& event) {
         if (event.type == sf::Event::Resized) {
-            updateLayout();
+            updateLayout(sf::Vector2u(event.size.width, event.size.height));
         }
         
         if (event.type == sf::Event::MouseButtonPressed) {
@@ -133,15 +138,11 @@ namespace rtype::client::gui {
                 }
                 // Check public servers button
                 else if (isPointInRect(mousePos, publicButtonRect)) {
-                    if (onPublicServersClick) {
-                        onPublicServersClick();
-                    }
+                    onPublicServersClick();
                 }
                 // Check private servers button
                 else if (isPointInRect(mousePos, privateButtonRect)) {
-                    if (onPrivateServersClick) {
-                        onPrivateServersClick();
-                    }
+                    onPrivateServersClick();
                 }
                 // Click outside - stop typing
                 else {
@@ -188,7 +189,7 @@ namespace rtype::client::gui {
         }
     }
     
-    void MainMenu::update(float deltaTime) {
+    void MainMenuState::update(float deltaTime) {
         // Cursor blinking animation
         cursorTimer += deltaTime;
         if (cursorTimer >= 0.5f) {
@@ -204,7 +205,10 @@ namespace rtype::client::gui {
         usernameText.setPosition(boxBounds.left + 10, boxBounds.top + 15);
     }
     
-    void MainMenu::render() {
+    void MainMenuState::render(sf::RenderWindow& window) {
+        // Update layout if needed
+        updateLayout(window.getSize());
+        
         // Render title
         window.draw(titleText);
         
@@ -223,20 +227,30 @@ namespace rtype::client::gui {
         window.draw(privateServersButton);
     }
     
-    void MainMenu::setPublicServersCallback(std::function<void()> callback) {
-        onPublicServersClick = callback;
+    void MainMenuState::onPublicServersClick() {
+        std::cout << "Switching to Public Server state" << std::endl;
+        std::string finalUsername = username.empty() ? generateUsername() : username;
+        stateManager.changeState(std::make_unique<PublicServerState>(stateManager, finalUsername));
     }
     
-    void MainMenu::setPrivateServersCallback(std::function<void()> callback) {
-        onPrivateServersClick = callback;
+    void MainMenuState::onPrivateServersClick() {
+        std::cout << "Switching to Private Server state" << std::endl;
+        std::string finalUsername = username.empty() ? generateUsername() : username;
+        stateManager.changeState(std::make_unique<PrivateServerState>(stateManager, finalUsername));
     }
     
-    bool MainMenu::isPointInRect(const sf::Vector2f& point, const sf::RectangleShape& rect) {
+    std::string MainMenuState::generateUsername() {
+        // Generate random 4-digit number between 1000-9999
+        int randomNum = 1000 + (rand() % 9000);
+        return "USERNAME" + std::to_string(randomNum);
+    }
+    
+    bool MainMenuState::isPointInRect(const sf::Vector2f& point, const sf::RectangleShape& rect) {
         sf::FloatRect bounds = rect.getGlobalBounds();
         return bounds.contains(point);
     }
     
-    void MainMenu::centerText(sf::Text& text, float x, float y) {
+    void MainMenuState::centerText(sf::Text& text, float x, float y) {
         sf::FloatRect textBounds = text.getLocalBounds();
         text.setPosition(x - textBounds.width / 2, y - textBounds.height / 2);
     }
