+++ client/src/gui/states/PrivateServerLobbyState.cpp
@@ -1,5 +1,5 @@
-#include "PrivateServerLobbyState.hpp"
-#include "MainMenuState.hpp"
+#include "gui/states/PrivateServerLobbyState.hpp"
+#include "gui/states/MainMenuState.hpp"
 #include <iostream>
 #include <cstdlib>
