--- client/src/render.cpp
@@ -1,6 +1,0 @@
-#include "components/render.hpp"
-
-namespace rtype::client::components {
-    Renderable::Renderable(const std::string& texturePath, float width, float height, RenderLayer layer)
-        : texturePath(texturePath), width(width), height(height), layer(layer), visible(true), alpha(1.0f) {}
-}
